refactor(vm): give the transitional UDS vsock adapter its own module
vsock.rs is already a section-divided monolith and the Firecracker PR is
about to move its handshake block out; the adapter that dies with the last
vsock_uds_path field should not grow it further. It lives in vsock/uds.rs
(a child module — the same path `vsock::UdsVsock` for callers) until R1
commit 17 deletes it; the remaining protocol clients get their own files
once the move in #658 has landed, so the two moves do not fight.
